test: redistribute coverage tests into feature-specific integration files
Move the remaining tests from the aggregated coverage.test.js /
final-coverage.test.js files into the feature-specific integration files
that already exist (alias.test.js, roots.test.js, pnp.test.js,
browserField.test.js, exportsField.test.js, importsField.test.js,
extension-alias.test.js, restrictions.test.js, description-file.test.js,
unsafe-cache.test.js, plugins.test.js, identifier.test.js). This mirrors
the pre-existing layout where each file focuses on a single feature of
the resolver.

The remaining cross-cutting integration cases live in
test/resolve-logging.test.js (directory/file existence logging, fragment
handling, LogInfoPlugin false-valued branches, top-level resolve sanity,
FileExistsPlugin "is not a file"). Pure helpers that have no integration
path continue to live in dedicated "unit-style" files matching the
pre-existing pattern (test/identifier.test.js, test/getPaths.test.js,
test/forEachBail.test.js): test/DescriptionFileUtils.test.js,
test/browser-shims.test.js, test/util-fs.test.js. The new tests added
earlier for util/path, util/strip-json-comments and util/entrypoints
keep the same shape as those existing utility tests.

Coverage held at 96.78% statements, 92.35% branches, 99.67% functions,
98.47% lines across 909 tests.
