+++ test/roots.test.js
@@ -135,3 +135,39 @@ describe("roots", () => {
 		);
 	});
 });
+
+describe("roots guard clauses", () => {
+	const fixtures = path.resolve(__dirname, "fixtures");
+	const fileSystem = new CachedInputFileSystem(fs, 4000);
+	const resolver = ResolverFactory.createResolver({
+		extensions: [".js"],
+		roots: [fixtures],
+		fileSystem,
+	});
+
+	it("falls through for empty requests (no root rewriting)", (done) => {
+		// Empty request should produce an error, not a root rewrite.
+		resolver.resolve({}, fixtures, "", {}, (err) => {
+			expect(err).toBeInstanceOf(Error);
+			done();
+		});
+	});
+
+	it("falls through for relative requests (no root rewriting)", (done) => {
+		// Relative request resolves normally (not via roots).
+		resolver.resolve({}, fixtures, "./a.js", {}, (err, result) => {
+			if (err) return done(err);
+			expect(result).toEqual(path.join(fixtures, "a.js"));
+			done();
+		});
+	});
+
+	it("rewrites absolute-style requests against roots", (done) => {
+		// Absolute request "/a.js" should be rewritten to roots[0] + /a.js
+		resolver.resolve({}, fixtures, "/a.js", {}, (err, result) => {
+			if (err) return done(err);
+			expect(result).toEqual(path.join(fixtures, "a.js"));
+			done();
+		});
+	});
+});
